Implement scrollable in Sectormaps
Becouse the default behavor with using a jpanel in jscrollpane is based on the height and width of the jpanel's child compoents and such (I barely undersand it myself), the viewport were behaving oddly, locking to the preferred size fo the client panel.

However, jScrolledPane has an alternate behavior when it comes to components implementing Scrollable. And with that, the viewport finally became adjustable.