Rework the game event system
RandomEvent is now an abstract class, with some properties now moved to functions.
The list of events is now a primitive Arrray, and not an ArrayList.
Events are now initiated in the list initialization, rather than added by variable names.
RandomEvent variable names has been kept to be used as a title for each event.
RandomEvent has been moved to its own file.
Notes on futher improvement has been added.