Move ShipLoc out of HomeShip
All the constant need for refrences to HomeShip.ShipLoc is causing a bit of clutter.
With the change of HomeShip becoming a singleton with static access, it is also no longer nessesary for ShipLoc to be an inner class.

Private HomeShip properties refrenced by ShipLoc are now set to protected access, and ShipLoc now uses a static refrence for those properties.